+++ allennlp_models/lm/models/bidirectional_lm.py
@@ -6,7 +6,7 @@ from allennlp.nn import InitializerApplicator
 
 from allennlp_models.lm.models.language_model import LanguageModel
 
-from allennlp_models.common.pretrained_model_config import PretrainedModelConfiguration
+from allennlp_models.common.model_card import add_pretrained_model
 
 
 @Model.register("bidirectional-language-model")
@@ -66,4 +66,4 @@ class BidirectionalLanguageModel(LanguageModel):
         )
 
 
-model_config = PretrainedModelConfiguration.from_dict({"id": "lm-bidirectional-language-model"})
+add_pretrained_model(id="lm-bidirectional-language-model")