+++ allennlp_models/lm/models/masked_language_model.py
@@ -11,7 +11,7 @@ from allennlp.nn import util, InitializerApplicator
 from allennlp.training.metrics import Perplexity
 from allennlp_models.lm.modules.language_model_heads import LanguageModelHead
 
-from allennlp_models.common.pretrained_model_config import PretrainedModelConfiguration
+from allennlp_models.common.model_card import add_pretrained_model
 
 
 @Model.register("masked_language_model")
@@ -186,10 +186,8 @@ class MaskedLanguageModel(Model):
     default_predictor = "masked_language_model"
 
 
-model_config = PretrainedModelConfiguration.from_dict(
-    {
-        "id": "lm-masked-language-model",
-        "name": "BERT-based Masked Language Model",
-        "archive_file": "bert-masked-lm-2019.09.17.tar.gz",
-    }
+add_pretrained_model(
+    id="lm-masked-language-model",
+    name="BERT-based Masked Language Model",
+    archive_file="bert-masked-lm-2019.09.17.tar.gz",
 )