+++ allennlp_models/lm/models/next_token_lm.py
@@ -11,7 +11,7 @@ from allennlp.nn import util, InitializerApplicator
 from allennlp.training.metrics import Perplexity
 from allennlp_models.lm.modules.language_model_heads import LanguageModelHead
 
-from allennlp_models.common.pretrained_model_config import PretrainedModelConfiguration
+from allennlp_models.common.model_card import add_pretrained_model
 
 
 @Model.register("next_token_lm")
@@ -150,10 +150,8 @@ class NextTokenLM(Model):
     default_predictor = "next_token_lm"
 
 
-model_config = PretrainedModelConfiguration.from_dict(
-    {
-        "id": "lm-next-token-lm-gpt2",
-        "name": "GPT2-based Next Token Language Model",
-        "archive_file": "gpt2-next-word-lm-2020.06.30.tar.gz",
-    }
+add_pretrained_model(
+    id="lm-next-token-lm-gpt2",
+    name="GPT2-based Next Token Language Model",
+    archive_file="gpt2-next-word-lm-2020.06.30.tar.gz",
 )