+++ allennlp_models/rc/models/qanet.py
@@ -18,8 +18,11 @@ from allennlp_models.rc.models.utils import (
 )
 from allennlp_models.rc.metrics import SquadEmAndF1
 
+from allennlp_models.common.model_card import add_pretrained_model
+
 
 @Model.register("qanet")
+@Model.register("rc-qanet")
 class QaNet(Model):
     """
     This class implements Adams Wei Yu's `QANet Model <https://openreview.net/forum?id=B14TlG-RW>`_
@@ -289,3 +292,6 @@ class QaNet(Model):
         }
 
     default_predictor = "reading_comprehension"
+
+
+add_pretrained_model(id="rc-qanet")