+++ allennlp_models/rc/models/transformer_qa.py
@@ -20,6 +20,8 @@ from allennlp_models.rc.models.utils import (
 )
 from allennlp_models.rc.metrics import SquadEmAndF1
 
+from allennlp_models.common.model_card import add_pretrained_model
+
 logger = logging.getLogger(__name__)
 
 
@@ -245,3 +247,6 @@ class TransformerQA(Model):
         }
 
     default_predictor = "transformer_qa"
+
+
+add_pretrained_model(id="rc-transformer-qa", name="Transformer QA")