+++ allennlp_models/structured_prediction/models/srl.py
@@ -19,6 +19,8 @@ from allennlp_models.structured_prediction.metrics.srl_eval_scorer import (
     DEFAULT_SRL_EVAL_PATH,
 )
 
+from allennlp_models.common.model_card import add_pretrained_model
+
 
 def write_bio_formatted_tags_to_file(
     prediction_file: TextIO,
@@ -160,6 +162,7 @@ def convert_bio_tags_to_conll_format(labels: List[str]):
 
 
 @Model.register("srl")
+@Model.register("sp-srl")
 class SemanticRoleLabeler(Model):
     """
     This model performs semantic role labeling using BIO tags using Propbank semantic roles.
@@ -489,3 +492,8 @@ def write_to_conll_eval_file(
     write_bio_formatted_tags_to_file(
         prediction_file, gold_file, verb_index, sentence, prediction, gold_labels
     )
+
+
+add_pretrained_model(
+    id="sp-srl", name="Open Information Extraction", archive_file="openie-model.2020.03.26.tar.gz"
+)