+++ allennlp_models/tagging/models/crf_tagger.py
@@ -14,6 +14,8 @@ from allennlp.nn import InitializerApplicator
 import allennlp.nn.util as util
 from allennlp.training.metrics import CategoricalAccuracy, SpanBasedF1Measure
 
+from allennlp_models.common.model_card import add_pretrained_model
+
 
 @Model.register("crf_tagger")
 class CrfTagger(Model):
@@ -299,3 +301,25 @@ class CrfTagger(Model):
         return metrics_to_return
 
     default_predictor = "sentence_tagger"
+
+
+add_pretrained_model(
+    id="tagging-elmo-crf-tagger",
+    name="ELMo-based Named Entity Recognition",
+    archive_file="ner-model-2020.02.10.tar.gz",
+    model_class=CrfTagger,
+)
+
+add_pretrained_model(
+    id="tagging-fine-grained-crf-tagger",
+    name="Fine Grained Named Entity Recognition",
+    archive_file="fine-grained-ner.2020-06-24.tar.gz",
+    model_class=CrfTagger,
+)
+
+add_pretrained_model(
+    id="tagging-fine-grained-transformer-crf-tagger",
+    name="Fine Grained Named Entity Recognition with Transformer",
+    archive_file="fgner_transformer.2020-07-14.tar.gz",
+    model_class=CrfTagger,
+)